pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "taf-saucedemo"
version = "0.1.0"
description = "UI Test Automation Framework on Playwright + Python for saucedemo.com"
readme = "README.md"
requires-python = ">=3.13"
license = { text = "MIT" }
authors = [ { name = "Raman Sharai" } ]

dependencies = [
  "playwright>=1.57.0",
  "pytest>=9.0.2",
  "pytest-html>=4.1.1",
  "allure-pytest>=2.15.3",
  "python-dotenv>=1.2.1",
  "typing-extensions>=4.15.0",
  "pytest-xdist>=3.8.0",
  "pytest-rerunfailures>=16.1",
  "pytest-cov>=7.0.0",
  "lxml>=6.0.2",
]

[project.optional-dependencies]
dev = [
  "black>=25.12.0",
  "flake8>=7.3.0",
  "flake8-pyproject>=1.2.4",
  "mypy>=1.19.1",
  "pre-commit>=4.5.1",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 120
target-version = ["py313"]

[tool.mypy]
python_version = "3.13"
warn_unused_configs = true
warn_return_any = true
warn_unused_ignores = true
strict_optional = true
no_implicit_optional = true
pretty = true
show_error_codes = true
exclude = ['tests/.*\.py$']

[tool.flake8]
max-line-length = 120
extend-ignore = ["E203", "W503"]
per-file-ignores = [
  "tests/*: D100,D101,D102,D103",
]
